refactor: extract shared Modal from QuizModal and SurveyModal
