--- src/components/QuizModal/styles.module.css
@@ -1,129 +1,0 @@
-.startButton {
-  background: var(--ifm-color-primary);
-  color: white;
-  border: none;
-  padding: 1rem 2rem;
-  font-size: 1.125rem;
-  font-weight: 600;
-  border-radius: 12px;
-  cursor: pointer;
-  transition: all 0.2s ease;
-  display: inline-flex;
-  align-items: center;
-  gap: 0.5rem;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-  margin: 2rem 0;
-}
-
-.startButton:hover {
-  transform: translateY(-2px);
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
-  opacity: 0.9;
-}
-
-.startButton:active {
-  transform: translateY(0);
-}
-
-.modalOverlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background: rgba(0, 0, 0, 0.75);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 9999;
-  padding: 1rem;
-  overflow-y: auto;
-  backdrop-filter: blur(4px);
-  animation: fadeIn 0.2s ease;
-}
-
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-
-.modalContent {
-  background: var(--ifm-background-color);
-  border-radius: 20px;
-  max-width: 800px;
-  width: 100%;
-  max-height: 90vh;
-  overflow-y: auto;
-  position: relative;
-  box-shadow: 0 20px 60px rgba(0, 0, 0, 0.3);
-  animation: slideUp 0.3s ease;
-}
-
-@keyframes slideUp {
-  from {
-    opacity: 0;
-    transform: translateY(20px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-.closeButton {
-  position: sticky;
-  top: 1rem;
-  right: 1rem;
-  margin-left: auto;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  background: var(--ifm-color-primary);
-  border: none;
-  cursor: pointer;
-  transition: all 0.2s ease;
-  z-index: 10;
-  float: right;
-  margin-bottom: -40px;
-}
-
-.closeButton:hover {
-  background: var(--ifm-color-primary-dark);
-  transform: rotate(90deg);
-}
-
-.closeButton svg {
-  width: 24px;
-  height: 24px;
-  color: white;
-}
-
-.quizWrapper {
-  padding: 1rem;
-  clear: both;
-}
-
-/* Mobile optimizations */
-@media (max-width: 768px) {
-  .modalContent {
-    max-height: 95vh;
-    border-radius: 16px 16px 0 0;
-    margin-top: auto;
-  }
-
-  .quizWrapper {
-    padding: 0.5rem;
-  }
-}
-
-/* Dark mode adjustments */
-html[data-theme='dark'] .modalOverlay {
-  background: rgba(0, 0, 0, 0.85);
-}
